"Updated class files in comp3607a2 package"
Final Build Alpha b

+++ src/main/java/comp3607a2/SpellingAttempt.java
@@ -43,7 +43,7 @@ public class SpellingAttempt {
     public String getDuration(){
         return duration;
     }
-    public double getSteps(){
+    public int getSteps(){
         return this.actions.size();
     }
     public int getErrorCorrections(){
